Refactor bower pluign
